nvim cmp tweaks

+++ config/nvim/lua/lsp.lua
@@ -112,7 +112,7 @@ local function global_on_attach(client, bufnr)
         buffer = bufnr,
         callback = function()
           if check_triggeredChars(triggerChars) then
-            vim.lsp.buf.signature_help { focus = false, silent = true, max_height = 7, border = "single" }
+            vim.lsp.buf.signature_help { focus = false, silent = true, max_height = 7, border = "double" }
           end
         end,
       })
